server/requirements.txt
python-dotenv
fastapi[all]
uvicorn
opencv-python>=4.12.0.88
misaki[en]>=0.8.2
mlx-lm==0.24.0
mlx-audio
pipecat-ai[openai,deepgram,rime,silero,mlx-whisper]>=0.0.76
aiortc
nltk
pipecat_ai_small_webrtc_prebuilt
pipecatcloud


